producto crud ok

+++ src/main/java/prueba/zemsania/demo/service/impl/productoServiceImp.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.sun.el.stream.Optional;
+
 import prueba.zemsania.demo.model.producto;
 import prueba.zemsania.demo.repository.productoRepository;
 import prueba.zemsania.demo.service.productoService;
@@ -24,7 +26,7 @@ public class productoServiceImp implements productoService {
 	@Override
 	public void saveProducto(producto product) throws Exception {
 		if(product.equals(getProducto(product.getIdProducto()))){
-			throw new Exception("Producto Existe");
+			throw new Exception("Este producto ya existe");
 		}else{
 			productRepository.save(product);
 		}
@@ -32,18 +34,43 @@ public class productoServiceImp implements productoService {
 	}
 
 	@Override
-	public void deleteProducto(int idProducto) {
-		productRepository.deleteById(idProducto);
+	public void deleteProducto(int idProducto) throws Exception {
+		try {
+            java.util.Optional<producto> producto = productRepository.findById(idProducto);
+            if (producto.isPresent()) {
+            	productRepository.deleteById(idProducto);
+            }
+        } catch (Exception ex) {
+            throw new Exception("Este producto no existe");
+        }
 	}
 
 	@Override
-	public producto getProducto(int idProducto) {
-		return productRepository.findById(idProducto).get();
+	public producto getProducto(int idProducto) throws Exception {
+		
+		try {
+            java.util.Optional<producto> producto = productRepository.findById(idProducto);
+            if (producto.isPresent()) {
+                return producto.get();
+            } else {
+                return null;
+            }
+
+        } catch (java.util.NoSuchElementException ex) {
+            throw new Exception("Este producto no existe");
+        }
 	}
 
 	@Override
-	public void updateProducto(producto product) {
-		productRepository.updateProduct(product.getIdProducto(), product.getNombre(), product.getPrecio());
+	public void updateProducto(producto product) throws Exception {
+		try {
+            java.util.Optional<producto> producto = productRepository.findById(product.getIdProducto());
+            if (producto.isPresent()) {
+            	productRepository.updateProduct(product.getIdProducto(), product.getNombre(), product.getPrecio());
+            }
+        } catch (java.util.NoSuchElementException ex) {
+            throw new Exception("Este producto no existe");
+        }		
 	}
